sentiment-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.115.12
pydantic==2.11.5
torch==2.7.1
transformers==4.52.4
uvicorn==0.34.3
slowapi==0.1.9
python-dotenv==1.1.0
